+++ src/main/java/com/github/cowwoc/docker/exception/ResourceNotFoundException.java
@@ -3,10 +3,12 @@ package com.github.cowwoc.docker.exception;
 import java.io.IOException;
 import java.io.Serial;
 
+import static com.github.cowwoc.requirements10.java.DefaultJavaValidators.requireThat;
+
 /**
- * Thrown if a referenced container does not exist.
+ * Thrown if a referenced resource does not exist.
  */
-public class ContainerNotFoundException extends IOException
+public class ResourceNotFoundException extends IOException
 {
 	@Serial
 	private static final long serialVersionUID = 0L;
@@ -16,7 +18,7 @@ public class ContainerNotFoundException extends IOException
 	 *
 	 * @param message an explanation of what went wrong
 	 */
-	public ContainerNotFoundException(String message)
+	public ResourceNotFoundException(String message)
 	{
 		super(message);
 	}
@@ -27,8 +29,9 @@ public class ContainerNotFoundException extends IOException
 	 * @param cause the underlying exception
 	 * @throws NullPointerException if {@code cause} is null
 	 */
-	public ContainerNotFoundException(Throwable cause)
+	public ResourceNotFoundException(Throwable cause)
 	{
 		super(cause);
+		requireThat(cause, "cause").isNotNull();
 	}
 }